+++ tests/sample/Test/Functional/transaction-api/CreateTxToSpendMultisigFundsWithBuilderEndpointTest.php
@@ -5,10 +5,10 @@ namespace sample\Test\Functional\transaction;
 use sample\Test\Functional\WebTestCase;
 
 /**
- * Class FundMultisignAddressTransactionTest
+ * Class CreateTxToSpendMultisigFundsWithBuilderEndpointTest
  * @package sample\Test\Functional\transaction
  */
-class FundMultisignAddressTransactionTest extends WebTestCase
+class CreateTxToSpendMultisigFundsWithBuilderEndpointTest extends WebTestCase
 {
     public function setUp()
     {
@@ -27,7 +27,7 @@ class FundMultisignAddressTransactionTest extends WebTestCase
         return join('', array_slice(explode('\\', get_class($this)), -1));
     }
 
-    public function testFundMultisignAddressTransaction()
+    public function testCreateTxToSpendMultisigFundsWithBuilderEndpoint()
     {
         $this->client->request('GET', $this->url);
         $responseBody = (string)$this->client->getResponse()->getContent();